gcc: Clean-up of non-C++0x compliant code, first steps
This patch cleans up a number of minor issues aiming to get closer to
compliance with the C++0x standard as interpreted by gcc and clang
(compile with std=c++0x and -pedantic-errors). In particular, the
patch cleans up enums where the last item was succeded by a comma,
namespaces closed by a curcly brace followed by a semi-colon, and the
use of the GNU-extension typeof (replaced by templated functions). It
does not address variable-length arrays, zero-size arrays, anonymous
structs, range expressions in switch statements, and the use of long
long. The generated CPU code also has a large number of issues that
remain to be fixed, mainly related to overflows in implicit constant
conversion (due to shifts).

+++ src/base/bitmap.cc
@@ -70,11 +70,13 @@ Bitmap::write(std::ostream *bmp) const
         // For further information see:
         //   http://en.wikipedia.org/wiki/BMP_file_format
         Magic magic = {{'B','M'}};
-        Header header = {sizeof(VideoConvert::Rgb8888) * width * height,
-                                0, 0, 54};
-        Info info = {sizeof(Info), width, height, 1,
-                    sizeof(VideoConvert::Rgb8888) * 8, 0,
-                    sizeof(VideoConvert::Rgb8888) * width * height, 1, 1, 0, 0};
+        Header header = {
+            static_cast<uint32_t>(sizeof(VideoConvert::Rgb8888)) *
+            width * height, 0, 0, 54};
+        Info info = {static_cast<uint32_t>(sizeof(Info)), width, height, 1,
+                     static_cast<uint32_t>(sizeof(VideoConvert::Rgb8888)) * 8,
+                     0, static_cast<uint32_t>(sizeof(VideoConvert::Rgb8888)) *
+                     width * height, 1, 1, 0, 0};
 
         char *p = headerBuffer = new char[sizeofHeaderBuffer];
         memcpy(p, &magic, sizeof(Magic));
